+++ Back end/package.json
@@ -1,12 +1,13 @@
 {
   "name": "waby",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "scripts": {
     "dev:server": "node --watch ./server/index.js"
   },
   "dependencies": {
+    "cors": "^2.8.5",
     "express": "4.19.2",
     "node": "^22.2.0",
     "react-dom": "^18.2.0",
